summary_dashboard/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
cryptography==3.4.8
Django==4.0
django-leaflet==0.28.3
django-location-field==2.1.0
django-static-fontawesome==5.15.4.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
huey==2.3.2
idna==2.10
numpy==1.24.1
pandas==1.5.2
psycopg2-binary==2.9.5
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
requests==2.25.1
six==1.16.0
sqlparse==0.4.3
tornado==6.2
urllib3==1.26.13
zope.event==4.6
zope.interface==6.0
